(PC-19704) feat(contentful adapteur): move each adaptor into is own file to enhance readability
