+++ src/libs/contentful/adapters/helpers/buildImageUrl.ts
@@ -1,0 +1,3 @@
+export const buildImageUrl = (url: string): string => {
+  return `https:${url}`
+}
