+++ src/libs/contentful/adapters/helpers/hasAtLeastOneField.ts
@@ -1,0 +1,3 @@
+export const hasAtLeastOneField = (object: any) => {
+  return Object.keys(object).length > 0
+}
