+++ src/libs/contentful/adapters/helpers/parseStringToNumber.ts
@@ -1,0 +1,3 @@
+export const parseStringToNumber = (offerId: string): number | undefined => {
+  return Number.isNaN(Number(offerId)) ? undefined : Number(offerId)
+}
