+++ src/libs/contentful/adapters/modules/adaptRecommendationModule.ts
@@ -1,0 +1,11 @@
+import { HomepageModuleType, RecommendedOffersModule } from 'features/home/types'
+import { RecommendationContentModel } from 'libs/contentful/types'
+
+export const adaptRecommendationModule = (
+  modules: RecommendationContentModel
+): RecommendedOffersModule => ({
+  type: HomepageModuleType.RecommendedOffersModule,
+  id: modules.sys.id,
+  displayParameters: modules.fields.displayParameters.fields,
+  recommendationParameters: modules.fields.recommendationParameters?.fields,
+})
